pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prereg-quality-llm"
version = "0.1.0"
description = "Research project evaluating preregistration quality using LLM scoring and human scoring"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai",
    "python-dotenv",
    "pyyaml",
    "jsonschema",
    "pydantic",
    "pandas",
    "numpy",
    "tqdm",
    "requests",
    "beautifulsoup4",
    "scikit-learn",
    "statsmodels",
    "lifelines",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

